refactor: move files in js template
