+++ templates/javascript/src/locations/ConfigScreen.jsx
