+++ templates/javascript/src/locations/ConfigScreen.spec.jsx
