+++ templates/javascript/src/locations/Dialog.jsx
