+++ templates/javascript/src/locations/Dialog.spec.jsx
