+++ templates/javascript/src/locations/EntryEditor.jsx
