+++ templates/javascript/src/locations/EntryEditor.spec.jsx
