+++ templates/javascript/src/locations/Page.jsx
