+++ templates/javascript/src/locations/Sidebar.jsx
