+++ templates/javascript/src/locations/Sidebar.spec.jsx
